+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.0.3
+You can now draw tiles!
+- Selecting an element in a dropdown in the Layers panel will automatically select the elements's layer. (Still haven't implemented anything other than tile drawing)
+- You can now see which element you have selected.
+- Drawing only affects the current layer.
+- The board updates after every modification.
+- Manually clicking on the "selected layer" radio buttons still does nothing, however. This will be implemented when I work on another layer.
+
 ## 0.0.2
 Dropdowns in the Layers panel are now functional. Nothing has been added yet.